+++ app/src/main/AndroidManifest.xml
@@ -33,20 +33,8 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".game.drawing.DrawingOffline">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
-
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
-        </activity>
-        <activity android:name=".game.drawing.DrawingOnline">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
-
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
-        </activity>
+        <activity android:name=".game.drawing.DrawingOffline"/>
+        <activity android:name=".game.drawing.DrawingOnline"/>
         <activity android:name=".auth.AccountCreationActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
