+++ app/src/main/java/ch/epfl/sweng/SDP/game/drawing/DrawingActivity.java
@@ -126,11 +126,4 @@ public class DrawingActivity extends Activity {
         }
     }
 
-    public void exitClick(View view) {
-        Log.d(TAG, "Exiting drawing view");
-        this.finish();
-    }
-
-
-
 }
